Pull a lot of changes and cruft cleanup from stuff I did for work.

+++ bootstrap.sh
@@ -1,31 +1,30 @@
 #!/bin/bash
 
-# OSX part commented out for two reasons:
-# a. I've borrowed it from a guy using yadm, but I desided to stick to homesick.
-# b. I have no Mac to test this on, but I'll want to have this ready as a future option.
+set -e
 
-#if [[ "$OSTYPE" == "darwin"* ]]; then
-#  echo "Doing OSX install of yadm"
-#
-#  # Brew can be fiddly, install separately:
-#  # ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
-#  hash brew 2>/dev/null || (echo "install brew first" && exit 1)
-#
-#  if hash yadm 2>/dev/null; then
-#    echo "yadm appears to be installed"
-#  else
-#    brew tap TheLocehiliosan/yadm
-#    brew update
-#    brew install git yadm
-#  fi
-#
-#  # Have ensured that yadm is available
-#  hash yadm 2>/dev/null || (echo "yadm install failed" && exit 1)
-#
-#  echo "getting some important extra brew packages"
-#  brew install screen neovim silversearcher-ag
-#
-#elif [[ "$(lsb_release -is)" == "Ubuntu" ]] || [[ "$(lsb_release -is)" == "Debian" ]] ; then
+if [[ "$OSTYPE" == "darwin"* ]]; then
+
+  ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+  hash brew 2>/dev/null || (echo "install brew first" && exit 1)
+
+  brew doctor
+  brew tap homebrew/versions
+  brew tap caskroom/versions
+  brew tap caskroom/fonts
+  brew tap chef/chef
+
+  echo "getting some important extra brew packages"
+  brew install homesick-completion screen neovim the_silver_searcher git curl
+  pip3 install homesick thefuck neovim powerline-status
+
+  echo "installing homesick"
+
+  sudo gem install homesick --no-ri --no-rdoc
+
+  # Have ensured that homesick is available
+  hash homesick 2>/dev/null || (echo "homesick install failed" && exit 1)
+
+elif [[ "$(lsb_release -is)" == "Ubuntu" ]] || [[ "$(lsb_release -is)" == "Debian" ]] ; then
   echo "Doing Debian install of homesick"
 
   echo "installing some essential packages"
@@ -66,8 +65,7 @@
       sudo update-alternatives --set vimdiff /usr/bin/vimdiff.nvim
     fi
   fi
-
-#fi
+fi
 
 ## Clone dotfiles
 homesick clone seefood/dotfiles dotfiles
@@ -81,24 +79,25 @@ yes | homesick symlink dotfiles
 mkdir -p ~/bin
 
 # Download Hub
-if [ ! -e "$HOME/bin/hub" ]; then
-  HUB_VERSION=2.3.0-pre10
+if [ ! -e "$HOME/bin/hub" ] && [ ! -e /usr/local/bin/hub ] ; then
+  HUB_VERSION=2.6.0
   HUB_DOWNLOAD_URL=https://github.com/github/hub/releases/download
   HUB_OS=hub-linux-amd64
 
   if [[ "$OSTYPE" == "darwin"* ]]; then
-    HUB_OS=hub-darwin-amd64
+    brew install hub
+  else
+
+    FULL_URL=${HUB_DOWNLOAD_URL}/v${HUB_VERSION}/${HUB_OS}-${HUB_VERSION}.tgz
+
+    echo "Downloading Hub ${HUB_VERSION}"
+
+    curl -fL ${FULL_URL} > /tmp/hub.tgz && \
+      tar zxf /tmp/hub.tgz -C /tmp && \
+      mv /tmp/${HUB_OS}-${HUB_VERSION}/bin/hub ~/bin/hub
+
+    rm -Rf /tmp/hub*
   fi
-
-  FULL_URL=${HUB_DOWNLOAD_URL}/v${HUB_VERSION}/${HUB_OS}-${HUB_VERSION}.tgz
-
-  echo "Downloading Hub ${HUB_VERSION}"
-
-  curl -fL ${FULL_URL} > /tmp/hub.tgz && \
-    tar zxf /tmp/hub.tgz -C /tmp && \
-    mv /tmp/${HUB_OS}-${HUB_VERSION}/bin/hub ~/bin/hub
-
-  rm -Rf /tmp/hub*
 else
   echo "Hub exists."
 fi
@@ -107,7 +106,7 @@ fi
 if hash fd; then
   echo "fd exists."
 else
-  fdversion=7.0.0
+  fdversion=7.2.0
   if [[ "$OSTYPE" == "darwin"* ]]; then
     brew install fd
   elif [[ "$(lsb_release -is)" == "Ubuntu" ]] || [[ "$(lsb_release -is)" == "Debian" ]] ; then
@@ -126,6 +125,12 @@ mkdir -p ~/.vim/autoload ~/.vim/bundle && \
 curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > /tmp/dein-installer.sh && \
     sh /tmp/dein-installer.sh ~/.vim/bundle
 
-git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf && ~/.fzf/install
+if [[ "$OSTYPE" == "darwin"* ]]; then
+  brew install fzf
+else
+  git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf && ~/.fzf/install
+fi
 
 ~/.homesick/repos/dotfiles/install_bash_it.sh
+~/.homesick/repos/dotfiles/install_general.sh
+~/.homesick/repos/dotfiles/install_dev.sh
